Support trace-level sampling
The libhoney-rust library has a `sample_rate` configuration option,
which provides random sampling of events sent to honeycomb.
Unfortunately, this sampling is done at the event level, rather than the
trace level. This means that spans or events belonging to a single trace
will have separate sampling decisions made for them. This in turn means
that when libhoney's sampling is enabled, traces will never have
complete span information within them. This severely limits the utility
of this sampling.

This proposed change introduces sampling based on the trace ID, meaning
that spans and events which are a part of the same trace will always
have the same sampling decision made for them. This will ensure that a
complete picture of the trace is sent up to honeycomb.

There are no breaking changes in behavior or API in this changeset, you
can still use the sampling logic built into libhoney if desired, but the
new function `new_honeycomb_telemetry_layer_with_trace_sampling` has
been introduced, which creates a telemetry layer that uses this new
sampling logic.

Alternatives:
- Use the `sample_rate` from the `libhoney` config struct as the
  trace-level sample rate, and reset the field to 1 before giving it to
  libhoney. This is undesirable as it makes the current behavior
  impossible to use, however it also makes it impossible to misuse this
  new function by providing a sample_rate to this function, and libhoney
  at the same time.
- Perform this trace-level sampling within `libhoney`. This is
  undesirable as it would require either parsing the traceID from the
  string provided to libhoney, or using a hash of this string. It is
  simpler to do the calculation in this library, where we have the raw
  trace ID. This would also be a breaking change for libhoney, whereas
  here it can be opt-in.