change: full rework of TraceId
Honeycomb allows it's trace_id's to be _any arbitrary string_. Respect that, because we may be sent an id from some source we do not directly control, such as an edge proxy.

This changes TraceId to avoid parsing, and to internally just hold a string. Unfortunately the tracing crate imposes restrictions that make this untenable for span ids.

This also includes a rework of sampling, making it be deterministic for a given trace_id, and also allowing it to function on any string.